improve file name, small refactoring
